test: Fix test failures when local postgres instance has a different table schema
Fixes an issue where tests failed if the local postgresql had a different schema.

For example, if I added a column to a table (but postgresql still has the old table schema),
the tests would fail (because we ran "CREATE TABLE IF NOT EXISTS" and "DELETE FROM", but at no point did we actually add the new column).

It's better to drop and recreate them entirely, I believe.